+++ pinax/stripe/webhooks/__init__.py
@@ -1,0 +1,4 @@
+from .base import Webhook  # noqa
+from .generated import *  # noqa
+from .overrides import CustomAccountApplicationDeauthorizeWebhook  # noqa
+from .registry import registry  # noqa
